partition `paramMatch` strongly typed
