+++ src/singleTable/model/definitions/partition/keySwap.spec.ts
@@ -1,17 +1,17 @@
-import { resolveKeySwaps, FullPartitionKeys } from './keySwap';
+import { resolveKeySwaps } from './keySwap';
 
 describe('key swaps tests', () => {
   it('should not change key getters if no param match is found', () => {
     const result = resolveKeySwaps({
-      getRangeKey: ({ someId }) => ['HI', someId],
+      getRangeKey: ({ someId }: { someId: string }) => ['HI', someId],
 
       getPartitionKey: () => ['PARTITION'],
-    }) as FullPartitionKeys<any, any, any>;
-
-    result.getKey({});
+    });
 
     expect(result.getPartitionKey()).toEqual(['PARTITION']);
     expect(result.getRangeKey({ someId: 'id!' })).toEqual(['HI', 'id!']);
+
+    // @ts-expect-error no unknown props
     expect(result.getKey({ someId: 'id!', other: 'should not be present' })).toEqual({
       partitionKey: ['PARTITION'],
       rangeKey: ['HI', 'id!'],
@@ -20,7 +20,7 @@ describe('key swaps tests', () => {
 
   it('should swap partition key param (single)', () => {
     const result = resolveKeySwaps({
-      getRangeKey: ({ someId }) => ['HI', someId],
+      getRangeKey: ({ someId }: { someId: string }) => ['HI', someId],
 
       getPartitionKey: ({ partId }) => ['PARTITION', partId],
 
@@ -33,6 +33,7 @@ describe('key swaps tests', () => {
     expect(result.getRangeKey({ someId: 'id!' })).toEqual(['HI', 'id!']);
 
     expect(
+      // @ts-expect-error we are testing paramMatchers, unknown or match props should be handled
       result.getKey({ someId: 'id!', other: 'should not be present', matchId: 'partSwap!' }),
     ).toEqual({
       partitionKey: ['PARTITION', 'partSwap!'],
@@ -42,7 +43,7 @@ describe('key swaps tests', () => {
 
   it('should swap range key param (single)', () => {
     const result = resolveKeySwaps({
-      getRangeKey: ({ someId }) => ['HI', someId],
+      getRangeKey: ({ someId }: { someId: string }) => ['HI', someId],
 
       getPartitionKey: () => ['PARTITION'],
 
@@ -52,9 +53,12 @@ describe('key swaps tests', () => {
     });
 
     expect(result.getPartitionKey()).toEqual(['PARTITION']);
+
+    // @ts-expect-error we are testing paramMatchers, unknown or match props should be handled
     expect(result.getRangeKey({ someId: 'id!', matchId: 'matchId!' })).toEqual(['HI', 'matchId!']);
 
     expect(
+      // @ts-expect-error we are testing paramMatchers, unknown or match props should be handled
       result.getKey({ someId: 'id!', matchId: 'matchId!', other: 'should not be present' }),
     ).toEqual({
       partitionKey: ['PARTITION'],
@@ -64,9 +68,21 @@ describe('key swaps tests', () => {
 
   it('should swap multiple key params', () => {
     const result = resolveKeySwaps({
-      getRangeKey: ({ someId, anotherId }) => ['HI', someId, anotherId],
+      getRangeKey: ({ someId, anotherId }: { someId: string; anotherId: string }) => [
+        'HI',
+        someId,
+        anotherId,
+      ],
 
-      getPartitionKey: ({ partId, keyId, timestamp }) => ['PARTITION', partId, keyId, timestamp],
+      getPartitionKey: ({
+        partId,
+        keyId,
+        timestamp,
+      }: {
+        partId: string;
+        keyId: string;
+        timestamp: string;
+      }) => ['PARTITION', partId, keyId, timestamp],
 
       paramMatch: {
         someId: 'matchId',
@@ -76,16 +92,19 @@ describe('key swaps tests', () => {
     });
 
     expect(
+      // @ts-expect-error we are testing paramMatchers
       result.getPartitionKey({ partId: 'partId', otherKeyId: 'otherKey', createdAt: 'some-tz' }),
     ).toEqual(['PARTITION', 'partId', 'otherKey', 'some-tz']);
 
     expect(
+      // @ts-expect-error we are testing paramMatchers
       result.getRangeKey({ someId: 'id!', matchId: 'matchId!', anotherId: 'another1' }),
     ).toEqual(['HI', 'matchId!', 'another1']);
 
     expect(
       result.getKey({
         someId: 'id!',
+        // @ts-expect-error we are testing paramMatchers
         matchId: 'matchId!',
         anotherId: 'another1',
         other: 'should not be present',
